requirements.txt
django
python-decouple
djangorestframework
markdown
drf-extensions